+++ Source.MaterialSystem/Source.MaterialSystem.csproj
@@ -16,4 +16,7 @@
 		<ProjectReference Include="..\Source.Common\Source.Common.csproj" />
 		<ProjectReference Include="..\Source.GUI.Controls\Source.GUI.Controls.csproj" />
 	</ItemGroup>
+	<ItemGroup>
+	  <Folder Include="Meshes\" />
+	</ItemGroup>
 </Project>
