--- Source.MaterialSystem/TempMeshGl46.cs
@@ -1,5 +1,0 @@
-﻿namespace Source.MaterialSystem;
-
-public class TempMeshGl46 : BaseMeshGl46 {
-
-}
